Move uniter resources params to apiserver/params
This is part of extracting the resources apiserver parts.
