+++ apiserver/params/resources_unit.go
@@ -1,16 +1,12 @@
-// Copyright 2016 Canonical Ltd.
+// Copyright 2017 Canonical Ltd.
 // Licensed under the AGPLv3, see LICENCE file for details.
 
-package private
-
-import "github.com/juju/juju/apiserver/params"
-
-// XXX move to params
+package params
 
 // ListResourcesArgs holds the arguments for an API request to list
 // resources for an application. The application is implicit to the uniter-
 // specific HTTP connection.
-type ListResourcesArgs struct {
+type ListUnitResourcesArgs struct {
 	// ResourceNames holds the names of the application's resources for
 	// which information should be provided.
 	ResourceNames []string `json:"resource-names"`
@@ -18,18 +14,18 @@ type ListResourcesArgs struct {
 
 // ResourcesResult holds the resource info for a list of requested
 // resources.
-type ResourcesResult struct {
-	params.ErrorResult
+type UnitResourcesResult struct {
+	ErrorResult
 
 	// Resources is the list of results for the requested resources,
 	// in the same order as requested.
-	Resources []ResourceResult `json:"resources"`
+	Resources []UnitResourceResult `json:"resources"`
 }
 
 // ResourceResult is the result for a single requested resource.
-type ResourceResult struct {
-	params.ErrorResult
+type UnitResourceResult struct {
+	ErrorResult
 
 	// Resource is the info for the requested resource.
-	Resource params.Resource `json:"resource"`
+	Resource Resource `json:"resource"`
 }
